Fixes a crash and a focussing issue
The crash would occur in some situations where StatsData could go invalid while waiting for a mutex.
The focussing issue would forward keystrokes to the game while having the focus in the console window. This is - for whatever reason - related to setting the locale to the english locale. Setting the english locale in the settings UI (where it is needed since imgui would not work properly with a different one) instead of at the end of the method which changed it, will not produce this issue.
It might be related to multithreading, but this change seems to "fix" it for now